+++ src/daemon/messages_test.go
@@ -69,7 +69,28 @@ func TestIntroductionMessage(t *testing.T) {
 			err: nil,
 		},
 		{
-			name: "INTR message with pubkey",
+			name: "INTR message with pubkey but empty user agent",
+			addr: "121.121.121.121:6000",
+			mockValue: daemonMockValue{
+				mirror:          10000,
+				protocolVersion: 1,
+				getMirrorPortResult: mirrorPortResult{
+					exist: false,
+				},
+				pubkey:           pubkey,
+				disconnectReason: ErrDisconnectInvalidUserAgent,
+			},
+			intro: &IntroductionMessage{
+				Mirror:  10001,
+				Port:    6000,
+				Version: 1,
+				valid:   false,
+				Extra:   append(pubkey[:], []byte{0, 0, 0, 0}...),
+			},
+			err: ErrDisconnectInvalidUserAgent,
+		},
+		{
+			name: "INTR message with pubkey and user agent",
 			addr: "121.121.121.121:6000",
 			mockValue: daemonMockValue{
 				mirror:          10000,
@@ -84,12 +105,12 @@ func TestIntroductionMessage(t *testing.T) {
 				Port:    6000,
 				Version: 1,
 				valid:   true,
-				Extra:   pubkey[:],
+				Extra:   append(pubkey[:], encoder.SerializeString("skycoin:0.24.1")...),
 			},
 			err: nil,
 		},
 		{
-			name: "INTR message with pubkey",
+			name: "INTR message with pubkey, user agent and additional data",
 			addr: "121.121.121.121:6000",
 			mockValue: daemonMockValue{
 				mirror:          10000,
@@ -104,27 +125,7 @@ func TestIntroductionMessage(t *testing.T) {
 				Port:    6000,
 				Version: 1,
 				valid:   true,
-				Extra:   pubkey[:],
-			},
-			err: nil,
-		},
-		{
-			name: "INTR message with pubkey and additional data",
-			addr: "121.121.121.121:6000",
-			mockValue: daemonMockValue{
-				mirror:          10000,
-				protocolVersion: 1,
-				getMirrorPortResult: mirrorPortResult{
-					exist: false,
-				},
-				pubkey: pubkey,
-			},
-			intro: &IntroductionMessage{
-				Mirror:  10001,
-				Port:    6000,
-				Version: 1,
-				valid:   true,
-				Extra:   append(pubkey[:], []byte("additional data")...),
+				Extra:   append(append(pubkey[:], encoder.SerializeString("skycoin:0.24.1")...), []byte("additional data")...),
 			},
 			err: nil,
 		},
@@ -144,8 +145,8 @@ func TestIntroductionMessage(t *testing.T) {
 				Mirror:  10001,
 				Port:    6000,
 				Version: 1,
-				valid:   true,
-				Extra:   pubkey2[:],
+				valid:   false,
+				Extra:   append(pubkey2[:], encoder.SerializeString("skycoin:0.24.1")...),
 			},
 			err: ErrDisconnectBlockchainPubkeyNotMatched,
 		},
@@ -165,10 +166,52 @@ func TestIntroductionMessage(t *testing.T) {
 				Mirror:  10001,
 				Port:    6000,
 				Version: 1,
-				valid:   true,
+				valid:   false,
 				Extra:   []byte("invalid extra data"),
 			},
 			err: ErrDisconnectInvalidExtraData,
+		},
+		{
+			name: "INTR message with pubkey, malformed user agent bytes",
+			addr: "121.121.121.121:6000",
+			mockValue: daemonMockValue{
+				mirror:          10000,
+				protocolVersion: 1,
+				getMirrorPortResult: mirrorPortResult{
+					exist: false,
+				},
+				pubkey:           pubkey,
+				disconnectReason: ErrDisconnectInvalidExtraData,
+			},
+			intro: &IntroductionMessage{
+				Mirror:  10001,
+				Port:    6000,
+				Version: 1,
+				valid:   false,
+				Extra:   append(pubkey[:], []byte{1, 2, 3}...),
+			},
+			err: ErrDisconnectInvalidExtraData,
+		},
+		{
+			name: "INTR message with pubkey, invalid user agent after parsing",
+			addr: "121.121.121.121:6000",
+			mockValue: daemonMockValue{
+				mirror:          10000,
+				protocolVersion: 1,
+				getMirrorPortResult: mirrorPortResult{
+					exist: false,
+				},
+				pubkey:           pubkey,
+				disconnectReason: ErrDisconnectInvalidUserAgent,
+			},
+			intro: &IntroductionMessage{
+				Mirror:  10001,
+				Port:    6000,
+				Version: 1,
+				valid:   false,
+				Extra:   append(pubkey[:], encoder.SerializeString("skycoin:0241")...),
+			},
+			err: ErrDisconnectInvalidUserAgent,
 		},
 		{
 			name: "Disconnect self connection",
@@ -212,7 +255,7 @@ func TestIntroductionMessage(t *testing.T) {
 			err: ErrDisconnectIncomprehensibleError,
 		},
 		{
-			name: "incomming connection",
+			name: "incoming connection",
 			addr: "121.121.121.121:12345",
 			mockValue: daemonMockValue{
 				mirror:                  10000,
@@ -266,6 +309,7 @@ func TestIntroductionMessage(t *testing.T) {
 			d.On("DaemonConfig").Return(DaemonConfig{
 				ProtocolVersion:    int32(tc.mockValue.protocolVersion),
 				MinProtocolVersion: int32(tc.mockValue.minProtocolVersion),
+				UserAgent:          "skycoin:0.24.1",
 			})
 			d.On("Mirror").Return(tc.mockValue.mirror)
 			d.On("IsDefaultConnection", tc.addr).Return(tc.mockValue.isDefaultConnection)